requirements/requirements.txt
Django==1.11
PyYAML==3.12
diff-match-patch==20110725.1
django-filter==1.1.0
django-import-export==1.0.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
et-xmlfile==1.0.1
jdcal==1.4
odfpy==1.3.6
openpyxl==2.5.3
pytz==2018.4
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
